+++ src/util/parser.h
@@ -22,7 +22,7 @@ public:
    */
   bool isInt(char *str)
   {
-    for (int i = 0; i < strlen(str); ++i)
+    for (size_t i = 0; i < strlen(str); ++i)
     {
       if (!isdigit(str[i]))
       {
